requirements.txt
cassandra-driver
requests
behave
kafka-python>=2.0.1
behave-restful>=0.3.1
assertpy>=1.0
jsonpath>=0.82
ConfigParser
rootpath>=0.1.1
retrying>=1.3.3
psycopg2-binary
